+++ 2-11-16 9_00PM/Assets/Scripts/ItemManager.cs
@@ -5,11 +5,11 @@ public class ItemManager : MonoBehaviour {
 
     public UnityEngine.UI.Text itemInfo; //Text showing information about the item
     private DataCollector data; // references the Click class
-    public float cost; // cost of the item
+    public double cost; // cost of the item
     public int tickValue; // the points/second value of that item
     public int count; // how many of this item you own
     public string itemName; // name of item
-    private float baseCost; // base cost of item
+    private double baseCost; // base cost of item
     public PointConverter converter;
     /// <summary>
     /// runs at start of game
@@ -39,14 +39,14 @@ public class ItemManager : MonoBehaviour {
         {
             data.setRadPoints(data.getRadPoints() - cost); //spends you points
             count++; //gives you one of this item
-            cost = Mathf.Round(baseCost * Mathf.Pow(1.15f, count)); //item increases in cost by 15%
+            cost = Mathf.Round((float)(baseCost * Mathf.Pow(1.15f, count))); //item increases in cost by 15%
             data.setPointsPerSec(data.getPointPerSec() + tickValue);
         }
     }
 
     public void UpgradePowerValue(ItemManager other)
     {
-        float num = 0;
+        double num = 0;
         if (data.getRadPoints() >= cost) // checks to see if you have enough points to buy it
         {
             num = data.getPointPerSec() - (other.count * other.tickValue);
@@ -55,7 +55,7 @@ public class ItemManager : MonoBehaviour {
             this.count++;
             other.tickValue *= 5;
             this.tickValue = other.tickValue * 5;
-            this.cost = Mathf.Round(this.baseCost * Mathf.Pow(11.5f, this.count)); // cost increases by 15% each time you purchase it, rounds cost to nearest int
+            this.cost = Mathf.Round((float)(this.baseCost * Mathf.Pow(11.5f, this.count))); // cost increases by 15% each time you purchase it, rounds cost to nearest int
             data.setPointsPerSec(data.getPointPerSec() + (other.count * other.tickValue));
         }
