+++ 2-11-16 9_00PM/Assets/Scripts/UpgradeManager.cs
@@ -5,11 +5,11 @@ public class UpgradeManager : MonoBehaviour {
 
     private DataCollector data;
     public UnityEngine.UI.Text itemInfo; //A bit of text showing information about the item (name,cost,power)
-    public float cost; // how many points an item costs
+    public double cost; // how many points an item costs
     public int count = 0; // how many of this item you have
     public int clickPower; // how many points/click this item adds
     public string itemName; //name of the item
-    private float baseCost; //base cost of the item
+    private double baseCost; //base cost of the item
     public PointConverter converter;
     /// <summary>
     /// runs at the beginning of the game
@@ -44,7 +44,7 @@ public class UpgradeManager : MonoBehaviour {
             data.setRadPoints(data.getRadPoints() - cost); //spends your points
             count++; //gives you 1 of this item 
             data.setPointsPerClick(data.getPointsPerClick() + clickPower);  // adds the power to the total points/click
-            cost = Mathf.Round(baseCost * Mathf.Pow(1.15f, count)); // cost increases by 15% each time you purchase it, rounds cost to nearest int
+            cost = Mathf.Round((float)(baseCost * Mathf.Pow(1.15f, count))); // cost increases by 15% each time you purchase it, rounds cost to nearest int
         }
     }
     /// <summary>
@@ -61,7 +61,7 @@ public class UpgradeManager : MonoBehaviour {
             this.count++;
             other.clickPower = clickPower;
             this.clickPower = (int)(clickPower * 1.5);
-            this.cost = Mathf.Round(this.baseCost * Mathf.Pow(11.5f, this.count)); // cost increases by 15% each time you purchase it, rounds cost to nearest int
+            this.cost = Mathf.Round((float)(this.baseCost * Mathf.Pow(11.5f, this.count))); // cost increases by 15% each time you purchase it, rounds cost to nearest int
             ClickPower(other.count,other.clickPower);
         }
